Node Side Script
This shall be used as the code for the node wherein the node shall send 1000 packets to the gateway. After, every packet, there is a gap of 2 seconds. Finally, the node comes into receiving mode to get the meta-data of all the 1000 packets. Moreover, this code is meant to save all the meta-data to into excel spreadsheet for plotting graphs and further.